+++ src/program-instructions.ts
@@ -1923,13 +1923,15 @@ export interface StateParams {
   pricingFrequencySeconds: number;
   liquidatorLiquidationPercentage: number;
   insuranceVaultLiquidationPercentage: number;
-  nativeTradeFeePercentage: anchor.BN;
-  nativeUnderlyingFeePercentage: anchor.BN;
+  nativeD1TradeFeePercentage: anchor.BN;
+  nativeD1UnderlyingFeePercentage: anchor.BN;
   nativeWhitelistUnderlyingFeePercentage: anchor.BN;
   nativeDepositLimit: anchor.BN;
   expirationThresholdSeconds: number;
   positionMovementFeeBps: number;
   marginConcessionPercentage: number;
+  nativeOptionTradeFeePercentage: anchor.BN;
+  nativeOptionUnderlyingFeePercentage: anchor.BN;
 }
 
 export interface UpdatePricingParametersArgs {